Revert "oops did not mean to change that"
This reverts commit 05572ff5dc25c9d5f6d29445174095d1c79753f2.

+++ src/r_main.c
@@ -475,7 +475,7 @@ angle_t R_PointToPseudoAngle(fixed_t x, fixed_t y)
 	}
 	else
 	{
-		double result = vecy / (vecx + vecy);
+		INT64 result = vecy / (vecx + vecy);
 		if (vecx < 0)
 		{
 			result = 2.0 - result;
